+++ src/main/java/org/deco/gachicoding/user/application/UserAuthenticationService.java
@@ -1,8 +1,14 @@
-package org.deco.gachicoding.service;
+package org.deco.gachicoding.user.application;
 
 import lombok.RequiredArgsConstructor;
 import org.deco.gachicoding.domain.auth.Auth;
 import org.deco.gachicoding.domain.auth.AuthRepository;
+import org.deco.gachicoding.service.MailService;
+import org.deco.gachicoding.user.domain.User;
+import org.deco.gachicoding.user.domain.repository.UserRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
@@ -12,7 +18,9 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
-public class AuthService {
+public class UserAuthenticationService implements UserDetailsService {
+
+    private final UserRepository userRepository;
     private final AuthRepository authRepository;
     private final MailService mailService;
 
@@ -40,5 +48,18 @@ public class AuthService {
     public Auth checkToken(UUID authToken) {
         Optional<Auth> auth = authRepository.findByAuthToken(authToken);
         return auth.orElseThrow(() -> new IllegalArgumentException("유효하지 않은 토큰입니다."));
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        User user = userRepository.findByUserEmail(email)
+                .orElseThrow(() -> new UsernameNotFoundException("Invalid Username"));
+
+//        List<GrantedAuthority> roles = new ArrayList<>();
+//        roles.add(new SimpleGrantedAuthority("ROLE_USER")); // 유저 롤을 다시 적용해야함.
+
+        UserDetails userDetails = new UserAuthenticationDto(user.getUserEmail(), user.getUserPassword());
+
+        return userDetails;
     }
 }
